moved evaluator to tools
